refactor(#8): write에서 learning으로 domain 이름 변경
