+++ app/src/main/java/com/ssafy/neegongnaegong/presentation/personal/PersonalContract.kt
@@ -1,7 +1,7 @@
 package com.ssafy.neegongnaegong.presentation.personal
 
 import com.ssafy.neegongnaegong.domain.model.personal.StudyRecord
-import com.ssafy.neegongnaegong.domain.model.write.Tag
+import com.ssafy.neegongnaegong.domain.model.learning.Tag
 import com.ssafy.neegongnaegong.presentation.base.UiEffect
 import com.ssafy.neegongnaegong.presentation.base.UiEvent
 import com.ssafy.neegongnaegong.presentation.base.UiState
