+++ app/src/main/java/com/ssafy/neegongnaegong/presentation/personal/edit/StudyRecordEditViewModel.kt
@@ -1,7 +1,7 @@
 package com.ssafy.neegongnaegong.presentation.personal.edit
 
 import com.ssafy.neegongnaegong.domain.data.TagData
-import com.ssafy.neegongnaegong.domain.model.write.Tag
+import com.ssafy.neegongnaegong.domain.model.learning.Tag
 import com.ssafy.neegongnaegong.presentation.base.BaseViewModel
 import com.ssafy.neegongnaegong.presentation.timer.write.StudyRecordWriteViewModel.Companion.MAX_TAG_LIMIT
 import dagger.hilt.android.lifecycle.HiltViewModel
